% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlepublicdata.r
\docType{data}
\name{country_slice}
\alias{country_slice}
\title{Some countries statistics}
\format{A data frame containing 13 observations.}
\source{
DSPL Google Code Page Downloads:
\url{https://developers.google.com/public-data/docs/tutorial}
}
\description{
This data set is one used in the DSPL Tutorial. Specifically, it contains
the population magnitudes at country level since 1960 to 1963.
}
\keyword{datasets}
