\name{gender_country_slice}
\docType{data}
\alias{gender_country_slice}
\title{Some Countries statistics at Gender level}
\description{
  This data set is one used in the DSPL Tutorial. Specifically, it contains the
  population magnitudes at country and gender level since 1960 to 1961.
}
\usage{data(gender_country_slice)}
\format{A data frame containing 13 observations.}
\source{
  DSPL Google Code Page Downloads: \url{https://code.google.com/p/dspl/downloads/detail?name=tutorial1.0.zip&can=2&q=}
}
\keyword{datasets}