\name{genders}
\docType{data}
\alias{genders}
\title{Genders example data set}
\description{
  This data set is one used in the DSPL Tutorial. Specifically, it contains the
  basic columns used to define a categorical dimentions such as gender.
}
\usage{data(genders)}
\format{A data frame containing 2 observations.}
\source{
  DSPL Google Code Page Downloads: \url{https://code.google.com/p/dspl/downloads/detail?name=tutorial1.0.zip&can=2&q=}
}
\keyword{datasets}