\name{genMoreInfo}
\alias{genMoreInfo}
\title{Generates a dataframe used to complement a DSPL bundle}
\description{
  Parsing \emph{csv}, \emph{tab} or \emph{xls(x)} files at a specific directory path, genMore info generates a dataframe used to complete a DSPL bundle with a more complete concepts definition including description, url, etc..
}
\usage{
  genMoreInfo(path, encoding=getOption("encoding"), ext='csv', output=NA,
              action='merge', dec=".")
}
\arguments{
  \item{path}{String. Path to the folder where the tables are saved.}
  \item{encoding}{The enconding of the files to be parsed.}
  \item{ext}{The extension of the files to be parsed. For more info check out the function \code{\link{dspl}}}
  \item{output}{If defined, the place where to save the dataframe as tab file. Otherwise it returns a data frame object.}
  \item{action}{Tells the function what to do if there's a copy of the file.}
}
\details{
  If no output is defiend the function returns a dataframe containing concepts
  as observations. Using this, the user may add more descripting info about concepts.
  
  In the case of defining a output, the function writes a tab file with the
  dataframe described above. The user may recycle this file writing ``append''
  in the \code{action} argument.
}
\value{
  If no \code{output} defined, \code{genMoreInfo} returns a dataframe with the following columns.
  \item{id}{XML id of the concept (autogenerated)}
  \item{label}{The label of the concept (autogenerated)}
  \item{description}{A breif description of the concept}
  \item{topic}{The topic of the concept}
  \item{url}{A URL for the concept where, for example, to get more info}
  \item{totalName}{A total name as specified by DSPL languge (works for dimentional concepts)}
  \item{pluralName}{A total name as specified by DSPL language (works for dimentional concepts)}
}
\references{
  Google Public Data Explorer:
    \url{http://publicdata.google.com} 
}
\author{George Vega Yon <george.vega@nodoschile.org>}

\seealso{
  See also \code{\link{dspl}}
}