% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlepublicdata.r
\docType{data}
\name{state_slice}
\alias{state_slice}
\title{Some US States statistics}
\format{A data frame containing 9 observations.}
\source{
DSPL Google Code Page Downloads:
\url{https://code.google.com/p/dspl/downloads/detail?name=tutorial1.0.zip&can=2&q=}
}
\description{
This data set is one used in the DSPL Tutorial. Specifically, it contains
the population magnitudes and unemployment rate at state level since 1960 to
1963.
}
\keyword{datasets}
