\name{countries}
\docType{data}
\alias{countries}
\title{World countries example data set}
\description{
  This data set is one used in the DSPL Tutorial. Specifically, it contains the
  basic columns used to define geographical dimentions, in this case, countries.
}
\usage{data(countries)}
\format{A data frame containing 5 observations.}
\source{
  DSPL Google Code Page Downloads: \url{https://code.google.com/p/dspl/downloads/detail?name=tutorial1.0.zip&can=2&q=}
}
\keyword{datasets}
