\name{dspl}
\alias{dspl}  
\title{Builds DSPL metadata file}
\description{
  Parsing \emph{csv}, \emph{tab} or \emph{xls(x)} files at a specific directory path, dspl generates a complete DSPL file.
  If an output string is specified, the function generates the complete ZIP (DSPL file plus csv files) ready to be uploaded to Google Public Data Explorer.
}
\usage{
  dspl(path, output=NA, replace=F, targetNamespace="", timeFormat="yyyy",
       lang=c("es","en"), name=NA, description=NA, url=NA, providerName=NA,
       providerURL=NA, extension="csv", encoding=getOption("encoding"),
       moreinfo=NA, dec=".")
}
\arguments{
  \item{path}{String. Path to the folder where the tables (csv|tab|xls) are at.}
  \item{output}{String, optional. Path to the output ZIP file.}
  \item{replace}{Logical. If \code{output} ZIP file is defined exists, \code{dspl} replaces it.}
  \item{timeFormat}{String. The corresponding time format of the collection. Should be specified accordingly to joda-time format. See the Details section for more information.}
  \item{lang}{A list of strings of the languages supported by the dataset. Could be only one.}
  \item{name}{List of strings. The name of the dataset as defined accordingly to the \code{lang} list.}
  \item{description}{List of strings. Description of the dataset. It also supports multiple description as the \code{name}}
  \item{url}{The corresponding URL for the dataset.}
  \item{providerNAme}{List of strings. The data provider name.}
  \item{providerURL}{List of strings. The data provider website url.}
  \item{extension}{The extension of the tables in the 'path' folder. Currently supports csv, tab, xls and xlsx files; the last two using \code{\link{XLConnect}} library.}
  \item{encoding}{The char encoding of the input tables. Currently ignored for microsoft excel files.}
  \item{moreinfo}{A special tab file generated by the function \code{\link{genMoreInfo}} that contains a dataframe of the dataset concepts with more specifications such as description, topic, url, etc.}
}
\details{
  If no output is defiend the function returns a list of class \code{dspl} that
  among its contents has a xml object; otherwise, if an output is defined, the
  results consists on two things, a ZIP file containing a allready bundle to be
  uploaded at \url{publicdata.google.com} (a collection of csv and a dspl file)
  and a message (character object).
  
  
}
\value{
  If no \code{output} defined, \code{dspl} returns list of \code{\link{class}}
  "\code{dspl}".   
  
  An object of class "\code{dspl}" is a list containing:
  \item{\code{dspl}}{A character printed version of the DSPL file.}
  \item{\code{concepts.by.table}}{A data frame object of concepts stored by table.}
  \item{\code{dimtabs}}{A data frame containing dimentional tables.}
  \item{\code{slices}}{A data frame of slices.}
  \item{\code{concepts}}{A data frame of concepts (all of them).}
  \item{\code{dimentions}}{A data frame of dimentional concepts.}
  \item{\code{statistics}}{A matrix of statistics.}
  
  otherwise the function will build a ZIP file as specified in the output containing
  the CSV and XML files.
}
\references{
  Google Public Data Explorer:
  \url{http://publicdata.google.com} 
}
\author{George Vega Yon <george.vega@nodoschile.org>}

\examples{
  demo(dspl)
}