% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlepublicdata.r
\docType{data}
\name{genders}
\alias{genders}
\title{Genders example data set}
\format{A data frame containing 2 observations.}
\source{
DSPL Google Code Page Downloads:
\url{https://developers.google.com/public-data/docs/tutorial}
}
\description{
This data set is one used in the DSPL Tutorial. Specifically, it contains
the basic columns used to define a categorical dimensions such as gender.
}
\keyword{datasets}
