\name{googlePublicData}
\alias{googlePublicData-package}
\alias{googlePublicData}
\docType{package}
\title{
  Working with Google Public Data Explorer DSPL Metadata Files
}
\description{
\code{googlePublicData} provides a collection of functions designed for working
with 'Google Public Data Explorer'. Automatically builds up the corresponding DSPL (XML)
metadata files and CSV files; compressing all the files and leaving them ready
to be published on the 'Public Data Explorer'.

Also includes several data structure verifiers in order to avoid surprises while loading your ZIP file to Public Data Explorer page.
  
Please visit the project home for more information and examples: \url{http://github.com/gvegayon/googlePublicData}.

For other projects you can visit the author's website: \url{http://www.its.caltech.edu/~gvegayon}
} 
\details{
  \tabular{ll}{
    Package: \tab googlePublicData\cr
    Type: \tab Package\cr
    Version: \tab 0.15.7.28\cr
    Date: \tab 2012-07-28\cr
    License: \tab MIT + file LICENSE\cr
  }
}
\author{George G. Vega Yon \email{g.vegayon@gmail.com}}
\references{
  \itemize{
    \item googlePublicData project site: \url{http://github.com/gvegayon/googlePublicData}
    \item Public Data Explorer site: \url{http://publicdata.google.com/}
    \item Public Data Explorer Developers site: \url{https://developers.google.com/public-data/}
    \item googleVis package: \url{http://code.google.com/p/google-motion-charts-with-r/}
  }
}
\examples{
  \dontrun{
    demo(dspl)
    }
}
\keyword{ package }
