% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlepublicdata.r
\docType{package}
\name{googlePublicData}
\alias{googlePublicData}
\alias{googlePublicData-package}
\alias{googlePublicData-package}
\title{Working with Google's Public Data Explorer DSPL Metadata Files}
\description{
\code{googlePublicData} package provides a collection of functions to set up
Google Public Data Explorer data visualization tool with your own data,
building automatically the corresponding DSPL (XML) metadata file jointly
with the CSV files. All zipped up and ready to be published at Public Data
Explorer.
}
\details{
Also includes several data structure verifiers in order to avoid surprises
while loading your ZIP file to Public Data Explorer page.

Please visit the project home for more information and examples:
\url{http://github.com/gvegayon/googlePublicData}.
}
\examples{

  \dontrun{
    demo(dspl)
    }

}
\references{
\itemize{
\item googlePublicData project site:
\url{http://github.com/gvegayon/googlePublicData}
\item Public Data Explorer
site: \url{http://publicdata.google.com/}
\item Public Data Explorer
Developers site: \url{https://developers.google.com/public-data/}
\item
googleVis package:
\url{https://github.com/mages/googleVis#googlevis} }
}
\author{
George G. Vega Yon
}
\keyword{package}
