\name{rdspl}
\alias{rdpls-package}
\alias{rdspl}
\docType{package}
\title{
  An R library to build Google's Public Data Explorer DSPL Metadata
}
\description{
  rdspl packages provides of a collection of functions to generate Google Public Data Explorer DSPL metadata files.
  
  It supports building dynamic graphs and adding attributes to edges and nodes.
  
  Please visit the project home for more information:
    \url{https://code.google.com/p/rdspl}.
} 
\details{
  \tabular{ll}{
    Package: \tab rdspl\cr
    Type: \tab Package\cr
    Version: \tab 0.12.04.19\cr
    Date: \tab 2012-04-19\cr
    License: \tab GPL version 2 or later\cr
  }
}
\author{
  George Vega Yon <george.vega@nodoschile.org>
}
\note{
  See the DSPL code site for more info:
    \url{https://developers.google.com/public-data/}
}
\references{
  \itemize{
    \item rdspl project site: \url{https://code.google.com/p/rdspl}
    \item Public Data Explorer site: \url{http://publicdata.google.com/}
    \item DSPL Code site: \url{http://code.google.com/p/dspl/}
  }
}
\examples{
    demo(dspl)
}
\keyword{ package }