\name{state_slice}
\docType{data}
\alias{state_slice}
\title{Some US States statistics}
\description{
  This data set is one used in the DSPL Tutorial. Specifically, it contains the
  population magnitudes and unemployment rate at state level since 1960 to 1963.
}
\usage{data(state_slice)}
\format{A data frame containing 9 observations.}
\source{
  DSPL Google Code Page Downloads: \url{https://code.google.com/p/dspl/downloads/detail?name=tutorial1.0.zip&can=2&q=}
}
\keyword{datasets}