\name{states}
\docType{data}
\alias{states}
\title{US states example data set}
\description{
  This data set is one used in the DSPL Tutorial. Specifically, it contains the
  basic columns used to define geographical dimentions, in this case, US Satates.
}
\usage{data(states)}
\format{A data frame containing 8 observations.}
\source{
  DSPL Google Code Page Downloads: \url{https://code.google.com/p/dspl/downloads/detail?name=tutorial1.0.zip&can=2&q=}
}
\keyword{datasets}