% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googlepublicdata.r
\docType{data}
\name{states}
\alias{states}
\title{US states example data set}
\format{A data frame containing 8 observations.}
\source{
DSPL Google Code Page Downloads:
\url{https://developers.google.com/public-data/docs/tutorial}
}
\description{
This data set is one used in the DSPL Tutorial. Specifically, it contains
the basic columns used to define geographical dimensions, in this case, US
States.
}
\keyword{datasets}
